delete UITextFieldDelegate and corrections
